Projects/create_your_own_chatbot/pyproject.toml
[project]
name = "create-your-own-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.2.8",
    "python-dotenv>=1.1.1",
]
